requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.2
crispy-bootstrap5==0.4
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.3
django-allauth==0.45.0
django-bootstrap-v5==1.0.4
django-crispy-forms==1.12.0
django-heroku==0.3.1
gunicorn==20.1.0
idna==3.2
oauthlib==3.1.1
Pillow==8.2.0
psycopg2==2.9.1
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.2.0
